+++ test_unstructured/partition/pptx/test_pptx.py
@@ -4,6 +4,7 @@
 
 import os
 import pathlib
+from tempfile import SpooledTemporaryFile
 
 import pptx
 import pytest
@@ -349,6 +350,21 @@ def test_partition_pptx_from_file_metadata_date(mocker: MockFixture):
             file=f,
         )
 
+    assert elements[0].metadata.last_modified is None
+
+
+def test_partition_pptx_from_file_explicit_get_metadata_date(mocker: MockFixture):
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-power-point-malformed.pptx")
+    mocked_last_modification_date = "2029-07-05T09:24:28"
+
+    mocker.patch(
+        "unstructured.partition.pptx.get_last_modified_date_from_file",
+        return_value=mocked_last_modification_date,
+    )
+
+    with open(filename, "rb") as f:
+        elements = partition_pptx(file=f, date_from_file_object=True)
+
     assert elements[0].metadata.last_modified == mocked_last_modification_date
 
 
@@ -366,6 +382,18 @@ def test_partition_pptx_from_file_with_custom_metadata_date(mocker: MockFixture)
         elements = partition_pptx(file=f, metadata_last_modified=expected_last_modification_date)
 
     assert elements[0].metadata.last_modified == expected_last_modification_date
+
+
+def test_partition_pptx_from_file_without_metadata_date():
+    """Test partition_pptx() with file that are not possible to get last modified date"""
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-power-point-malformed.pptx")
+    with open(filename, "rb") as f:
+        sf = SpooledTemporaryFile()
+        sf.write(f.read())
+        sf.seek(0)
+        elements = partition_pptx(file=sf, date_from_file_object=True)
+
+    assert elements[0].metadata.last_modified is None
 
 
 def test_partition_pptx_element_metadata_has_languages():